new upload image handler with following enhancements.
First, by using (s/map-of keyword? multipart/temp-file-part), we allow any number of files to be posted to this endpoint. Unfortunately, this doesn’t play well with swagger, but we’ll be able to test it from our application shortly.
Second, instead of explicitly handling :avatar and :banner, we’re using reduce-kv to perform our image saving logic for each file in our multipart parameter map.
Third, we’ve added a cond statement with several checks in place. We reject files which are larger than 5 MiB, or which have a Content-Type that doesn’t start with image/. It is possible to add additional clauses to handle other media, such as video or audio, but we’ll leave that for now.